Server handles two commands, push and dc.
With push you'll send data to the server (the server will cache this)
With dc you'll execute the server code the write the cached data (sent earlier with the push command) to a file.

TODO: Create a random file name. Currently every time the DC command is called, the file will be overwritten (#Dataloss?)